+++ examples/serverless-otel/packages/functions/src/tracing.ts
@@ -12,7 +12,7 @@ new BaselimeSDK({
         captureHeaders: true,
       })
     ],
-    resourceDetectors: []
+    resourceDetectors: [],
   }).start();
 
 export {