+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [Unreleased]
+
+* add support for custom attributes
+  
 ## [0.4.7] - 2024-03-18
 
 * add support for passing custom detectors